Split out http&websocket example loops
